fix: refactored controllers into folders
