+++ app/Http/Controllers/Products/CategoryController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Products;
 
+use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
 
@@ -14,21 +15,6 @@ class CategoryController extends Controller
     }
 
     public function store(Request $request)
-    {
-        //
-    }
-
-    public function show(Category $category)
-    {
-        //
-    }
-
-    public function update(Request $request, Category $category)
-    {
-        //
-    }
-
-    public function destroy(Category $category)
     {
         //
     }
