+++ app/Http/Controllers/Users/AddressController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Users;
 
+use App\Http\Controllers\Controller;
 use App\Models\Address;
 use Illuminate\Http\Request;
