+++ routes/api.php
@@ -1,9 +1,9 @@
 <?php
 
-use App\Http\Controllers\CategoryController;
-use App\Http\Controllers\ProductController;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+
+use App\Http\Controllers\Products\CategoryController;
+use App\Http\Controllers\Products\ProductController;
 use App\Http\Controllers\Users\SessionController;
 use App\Http\Controllers\Users\UserController;
